Split platform-specific functions (2)
